refactor(audiotap): extract debug logging helpers to keep AppAudioCapture under 600 lines
Problem: After the Swift 6 strict concurrency migration on main (3c1c540)
added 9 lines (an @unchecked Sendable doc-block + the conformance) to
AppAudioCapture.swift, the merge with this branch puts the file at 607
lines — over the 600-line lint cap. CI's `swiftlint lint --strict`
fails with `file_length` and cascades cancellation to every other
job via `cancel-on-failure`. Locally the branch was at 600 lines
because main hadn't been picked up yet, so the failure only surfaced
on the GitHub-side merge.

Decision: Extract the three `publishCurrentLevel` / `accumulateDebugRMS`
/ `maybeReportDebugRMS` helpers into a sibling extension file
(`AppAudioCapture+DebugLogging.swift`). Same pattern as the existing
`+PIDTranslation` and `+LiveSink` extensions, and the same precedent
PR #307 set when an earlier inline migration pushed the file to 617
lines. The four backing properties (`debugLogging`, `debugRMS`,
`debugTotalBytes`, `levelPublisher`) widen from `private` to
`internal` — module-scoped only, not test-visibility-widening.

Rejected: `// swiftlint:disable file_length` would contradict CLAUDE.md
and just defer the same conversation. Trimming the upstream Swift 6
doc-block would land in someone else's PR.

Result: AppAudioCapture.swift 607 → 572 lines. swiftlint lint --strict
+ swiftlint analyze --strict both green locally. AudioTapLib 109-test
suite still passes.
