+++ tools/audiotap/Sources/AppAudioCapture.swift
@@ -24,7 +24,9 @@ public class AppAudioCapture: @unchecked Sendable {
     let sampleRate: Int
     private let channels: Int
     private let outputFileDescriptor: Int32
-    private let debugLogging: Bool
+    /// `internal` (not `private`) so the cross-file `+DebugLogging` extension
+    /// can drive the throttled dBFS log line.
+    let debugLogging: Bool
     let liveSink: LiveAudioSink?
     private var aggregateID = AudioObjectID(kAudioObjectUnknown)
     private var tapID = AudioObjectID(kAudioObjectUnknown)
@@ -37,9 +39,11 @@ public class AppAudioCapture: @unchecked Sendable {
         label: "audiotap.writer", qos: .userInteractive,
     )
 
-    private var debugRMS = DebugRMSReporter()
-    private var debugTotalBytes: UInt64 = 0
-    private let levelPublisher = LevelPublisher()
+    /// `internal` (not `private`) so the cross-file `+DebugLogging` extension
+    /// can drive the per-buffer RMS accumulator + dBFS report cadence.
+    var debugRMS = DebugRMSReporter()
+    var debugTotalBytes: UInt64 = 0
+    let levelPublisher = LevelPublisher()
 
     /// Returns the instantaneous app-audio level in dBFS, decayed to -120 if
     /// no buffer arrived in the last 0.5 seconds (e.g. tap died, device
@@ -563,45 +567,5 @@ extension AppAudioCapture {
         DispatchQueue.main.asyncAfter(deadline: .now() + delay) { [weak self] in
             self?.completeRestart()
         }
-    }
-}
-
-// MARK: - Level + debug logging helpers
-
-@available(macOS 14.2, *)
-private extension AppAudioCapture {
-    /// Publish the most recent per-buffer dBFS reading so UI consumers
-    /// (menu bar level indicator) can poll it. Called from the IOProc
-    /// after `accumulateDebugRMS`.
-    func publishCurrentLevel() {
-        levelPublisher.publish(level: debugRMS.lastLevelDBFS)
-    }
-
-    /// Sums squares of the interleaved Float32 buffer into the shared RMS reporter.
-    /// Called unconditionally from the IOProc; the dBFS log line is gated separately.
-    func accumulateDebugRMS(data: UnsafeMutableRawPointer, byteCount: Int) {
-        let count = byteCount / MemoryLayout<Float>.size
-        guard count > 0 else { return }
-        let buf = UnsafeBufferPointer(
-            start: data.assumingMemoryBound(to: Float.self), count: count,
-        )
-        var sumSq: Double = 0
-        for sample in buf {
-            sumSq += Double(sample) * Double(sample)
-        }
-        debugRMS.add(sumSq: sumSq, samples: count)
-        debugTotalBytes += UInt64(byteCount)
-    }
-
-    /// Drain the 5-s throttle and emit one RMS-energy log line per tick, but
-    /// only when `debugLogging` is on. The drain itself runs unconditionally
-    /// so the reporter's accumulators stay bounded for long sessions.
-    func maybeReportDebugRMS() {
-        guard let report = debugRMS.tick() else { return }
-        guard debugLogging else { return }
-        let dBStr = String(format: "%.1f", report.dBFS)
-        logger.info(
-            "[debug] App audio RMS (5s): \(dBStr, privacy: .public) dBFS, samples=\(report.samples, privacy: .public), totalBytes=\(self.debugTotalBytes, privacy: .public)",
-        )
     }
 }
